pyproject.toml
[tool.poetry]
name = "homecomfort"
version = "0.1.0"
description = ""
authors = ["Notebook <zekserj@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
mysqlclient = "^2.2.4"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
pillow = "^10.3.0"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
django-dotenv = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
